requirements.txt
streamlit
tensorflow
tf-keras
keras
transformers
numpy
Pillow
opencv-python
